backend/requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==2.7.0
eventlet==0.39.1
Flask==2.0.1
Flask-Caching==1.10.1
Flask-SocketIO==5.5.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==24.2
python-dotenv==0.18.0
python-engineio==4.11.2
python-socketio==5.12.1
selenium==3.141.0
setuptools==76.0.0
simple-websocket==1.1.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wsproto==1.2.0
zipp==3.4.1
